GH-10 Fix unit tests
