+++ src/test/java/org/springframework/samples/petclinic/service/userService/UserServiceSpringDataJpaTests.java
@@ -1,12 +1,53 @@
 package org.springframework.samples.petclinic.service.userService;
 
+import org.junit.Before;
+import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.mockito.MockitoAnnotations;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+
+import org.springframework.samples.petclinic.model.User;
+import org.springframework.samples.petclinic.repository.UserRepository;
+import org.springframework.samples.petclinic.service.UserService;
+
+import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.mockito.BDDMockito.verify;
 
 @SpringBootTest
 @RunWith(SpringJUnit4ClassRunner.class)
 @ActiveProfiles("spring-data-jpa, hsqldb")
-public class UserServiceSpringDataJpaTests extends AbstractUserServiceTests
-{}
+public class UserServiceSpringDataJpaTests
+{
+    @Autowired
+    private UserService userService;
+
+    @MockBean
+    private UserRepository userRepository;
+
+    @Before
+    public void init()
+    {
+        MockitoAnnotations.initMocks(this);
+    }
+
+    @Test
+    public void shouldAddUser() throws Exception
+    {
+        User user = new User();
+        user.setUsername("username");
+        user.setPassword("password");
+        user.setEnabled(true);
+        user.addRole("OWNER_ADMIN");
+
+        userService.saveUser(user);
+        assertThat(user.getRoles().parallelStream().allMatch(role -> role.getName().startsWith("ROLE_")), is(true));
+        assertThat(user.getRoles().parallelStream().allMatch(role -> role.getUser() != null), is(true));
+
+        verify(userRepository).save(user);
+    }
+}
